patch for $rm_unknown and $rm_storage_unknown sub
Patch for $rm_unknown and $rm_storage_unknown sub.

For $rm_unknown patch:
I was getting uninitialized value errors with ESXi 7.x on HPE ProLiant servers. So I did my best to figure out what was causing this problem. It turned out that the hash keys for different sensors have different state names.
So I added a check to determine which key is present. Then the status is checked. If the status is unknown, nothing is returned. Otherwise, the status is reported.

For $rm_storage_unknown patch:
It is a workaround to exclude some false positives from the result. The bug is related to ESXi 7.x and HPE ProLiant 10th generation servers that have an up-to-date SPP installed. The problem is a bug in the HPE RAID driver.
The ESXi reports random unconfigured disks. So I excluded them from the report.